+++ src/commands/account/generate/index.ts
@@ -3,7 +3,7 @@ import fs from 'fs'
 import path from 'path'
 import type { Hex } from 'viem'
 
-import { accounts } from '@classes'
+import { accounts, config } from '@classes'
 import { DATA_FOLDER } from '@classes/Config/Config'
 import { chalkInfo } from '@constants/chalk'
 
@@ -12,10 +12,13 @@ import { ensureAccessibilityAtPath } from '@utils/files'
 
 import { passwordWithConfirmPrompt } from '@commands/account/prompt'
 import { customErrorLog } from '@errors/log'
+import { sendOffChainEvent } from '@utils/analytics'
 import { getProjectChainInfo } from '@utils/config'
 
 export const handleGenerateAccount = async (accountId?: string) => {
   try {
+    const projectConfig = config.getProjectConfig()
+
     // 1. prompt account ID
     const validAccountId = await notDuplicatedAccountIdPrompt(accountId)
 
@@ -45,6 +48,10 @@ export const handleGenerateAccount = async (accountId?: string) => {
     console.log(
       `Please transfer ${chainInfo.nativeCurrency.symbol} to the address ${chalkInfo(walletAddress)} account on ${chalkInfo(chainInfo.name)} blockchain to cover gas fees.`,
     )
+
+    if (projectConfig.allow_analytics) {
+      sendOffChainEvent({ action: 'account generate', payload: { walletAddress } })
+    }
   } catch (error) {
     customErrorLog(error as Error)
   }