+++ src/commands/action/list/index.ts
@@ -3,6 +3,7 @@ import { protocolModules } from '@constants/protocol-module'
 import type { PROTOCOL_MODULE } from '@enums/module'
 import { ValidateInputValueError } from '@errors/validate'
 import type { InfinitActionRecord } from '@infinit-xyz/core'
+import { sendOffChainEvent } from '@utils/analytics'
 import Table from 'cli-table3'
 
 export const handleListAction = () => {
@@ -27,5 +28,10 @@ export const handleListAction = () => {
   }
 
   console.log(table.toString())
+
+  if (projectConfig.allow_analytics) {
+    sendOffChainEvent({ action: 'action list', payload: {} })
+  }
+
   return table
 }