+++ src/commands/script/generate/index.ts
@@ -7,6 +7,7 @@ import { protocolModules } from '@constants/protocol-module'
 import type { PROTOCOL_MODULE } from '@enums/module'
 import { ERROR_MESSAGE_RECORD } from '@errors/errorList'
 import { ValidateInputValueError } from '@errors/validate'
+import { sendOffChainEvent } from '@utils/analytics'
 import { ensureCwdRootProject } from '@utils/files'
 import type { ProtocolModuleActionKey } from './index.type'
 import { getScriptFileDirectory, getUniqueScriptFileName, handleGenerateScriptFile } from './utils'
@@ -14,8 +15,9 @@ import { getScriptFileDirectory, getUniqueScriptFileName, handleGenerateScriptFi
 export const handleGenerateScript = async (actionId?: string) => {
   ensureCwdRootProject()
 
+  const projectConfig = config.getProjectConfig()
   // Validate config
-  const protocolModule = protocolModules[config.getProjectConfig().protocol_module as PROTOCOL_MODULE]
+  const protocolModule = protocolModules[projectConfig.protocol_module as PROTOCOL_MODULE]
 
   if (!protocolModule) {
     throw new ValidateInputValueError(ERROR_MESSAGE_RECORD.PROTOCOL_NOT_SUPPORTED)
@@ -64,4 +66,8 @@ export const handleGenerateScript = async (actionId?: string) => {
   console.log(`⚙ Generate ${chalkInfo(selectedAction.name)} script successfully`)
   console.log()
   console.log(`📁 Checkout ${chalkInfo('src/scripts/' + scriptFileName + '.script.ts')}`)
+
+  if (projectConfig.allow_analytics) {
+    sendOffChainEvent({ action: 'script generate', payload: { action: actionKey, module: projectConfig.protocol_module } })
+  }
 }