+++ src/enums/chain.ts
@@ -1,9 +1,11 @@
 export enum CHAIN_ID {
   Ethereum = '1',
   Mantle = '5000',
-  BSC = '56',
-  Arbitrum = '42161',
-  ARB_SEPOLIA = '421614',
+  Sepolia = '11155111',
+  Holesky = '17000',
+  // BSC = '56',
+  // Arbitrum = '42161',
+  // ARB_SEPOLIA = '421614',
   // Optimism = '10',
   // Avalanche = '43114',
 }