+++ src/schemas/generated.d.ts
@@ -19,7 +19,14 @@ export interface InfinitConfigSchema {
   protocol_module: string
   chain_info: {
     name: string
+    short_name: string
     network_id: number
+    native_currency: {
+      name: string
+      symbol: string
+      decimals: number
+    }
     rpc_url: Url
   }
+  allow_analytics?: boolean
 }