+++ src/utils/project.ts
@@ -13,10 +13,12 @@ import type { PACKAGE_MANAGER } from '@enums/package-managers'
 import { ValidateInputValueError } from '@errors/validate'
 import { ContractProvider } from '@infinit-xyz/core'
 import type { InfinitConfigSchema } from '@schemas/generated'
+import { sendOffChainEvent } from '@utils/analytics'
 import { spawnChild } from '@utils/childprocess'
 import { writeFileSync } from '@utils/files'
 import { getPackageManagerInstallArgs } from '@utils/packageManager'
 import yaml from 'js-yaml'
+import _ from 'lodash' // [TODO/INVESTIGATE] later on importing from lodash
 import ora from 'ora'
 import os from 'os'
 import { dependencies as packageJsonDependencies, name as packageJsonName } from 'package.json'
@@ -73,7 +75,8 @@ export const initializeCliProject = async (
   chainId: CHAIN_ID,
   packageManager: PACKAGE_MANAGER,
   deployerId?: string,
-) => {
+  allowAnalytics?: boolean,
+): Promise<{ generatedScriptFile: string }> => {
   const packageJsonFile = fs.readFileSync(path.join(projectDirectory, 'package.json'), 'utf-8')
   const projectName = JSON.parse(packageJsonFile).name
 
@@ -107,9 +110,12 @@ export const initializeCliProject = async (
     protocol_module: protocolModule,
     chain_info: {
       name: chainInfo.name,
+      short_name: chainInfo.shortName,
       network_id: parseInt(chainInfo.chainId),
+      native_currency: chainInfo.nativeCurrency,
       rpc_url: chainInfo.rpcList[0],
     },
+    allow_analytics: allowAnalytics,
   }
   fs.writeFileSync(path.join(projectDirectory, 'src', FILE_NAMES.CONFIG), `# yaml-language-server: $schema=${CONFIG_SCHEMA}\n\n` + yaml.dump(config) + os.EOL)
 
@@ -121,7 +127,8 @@ export const initializeCliProject = async (
     throw new ValidateInputValueError('Initialize action not found')
   }
   const scriptFolderPath = getScriptFileDirectory(projectDirectory)
-  await handleGenerateScriptFile(initAction, DEFAULT_ACTION_KEY, protocolModuleInfo.libPath, 'default', scriptFolderPath, deployerId)
+  const camelCaseActionName = _.camelCase(initAction.actionClassName)
+  await handleGenerateScriptFile(initAction, DEFAULT_ACTION_KEY, protocolModuleInfo.libPath, camelCaseActionName, scriptFolderPath, deployerId)
 
   // Install dependencies
   console.log(`\n🚀 Initialize a project from ${chalkInfo(protocolModule)} module.`)
@@ -144,5 +151,13 @@ export const initializeCliProject = async (
 
   spinner.stopAndPersist({ symbol: '✅', text: `Dependencies installed successfully.` })
 
+  if (allowAnalytics) {
+    sendOffChainEvent({ action: 'init', payload: { project_name: projectName, protocol_module: protocolModule, chain_id: chainId } })
+  }
+
   spinner.stop()
+
+  return {
+    generatedScriptFile: `${camelCaseActionName}.script.ts`,
+  }
 }